+++ backend/School_Medical_System/src/main/java/sms/swp391/models/dtos/responses/StudentResponse.java
@@ -1,6 +1,8 @@
 package sms.swp391.models.dtos.responses;
 
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -24,6 +26,6 @@ public class StudentResponse {
     private String currentMedications;
     private String chronicDiseases;
     private String allergies;
-    private Float height;
-    private Float weight;
+    private BigDecimal height;
+    private BigDecimal weight;
 }
